requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
boto3==1.34.65
botocore==1.34.65
cffi==1.16.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
exceptiongroup==1.2.0
fastapi==0.110.0
filelock==3.13.1
fonttools==4.50.0
fsspec==2024.3.1
h11==0.14.0
idna==3.6
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jwt==1.3.1
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pillow==10.2.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
s3transfer==0.10.1
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
sns==0.1
starlette==0.36.3
sympy==1.12
threadpoolctl==3.3.0
torch==2.2.1
torchvision==0.17.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.28.1
